+++ Pod/Classes/KolodaView/KolodaView.swift
@@ -23,15 +23,15 @@ private let defaultAlphaValueSemiTransparent: CGFloat = 0.7
 
 public protocol KolodaViewDataSource:class {
     
-    func kolodaNumberOfCards(_ koloda: KolodaView) -> UInt
-    func koloda(_ koloda: KolodaView, viewForCardAtIndex index: UInt) -> UIView
-    func koloda(_ koloda: KolodaView, viewForCardOverlayAtIndex index: UInt) -> OverlayView?
+    func kolodaNumberOfCards(_ koloda: KolodaView) -> Int
+    func koloda(_ koloda: KolodaView, viewForCardAtIndex index: Int) -> UIView
+    func koloda(_ koloda: KolodaView, viewForCardOverlayAtIndex index: Int) -> OverlayView?
     
 }
 
 public extension KolodaViewDataSource {
     
-    func koloda(_ koloda: KolodaView, viewForCardOverlayAtIndex index: UInt) -> OverlayView? {
+    func koloda(_ koloda: KolodaView, viewForCardOverlayAtIndex index: Int) -> OverlayView? {
         return nil
     }
     
@@ -39,37 +39,37 @@ public extension KolodaViewDataSource {
 
 public protocol KolodaViewDelegate:class {
     
-    func koloda(_ koloda: KolodaView, allowedDirectionsForIndex index: UInt) -> [SwipeResultDirection]
-    func koloda(_ koloda: KolodaView, shouldSwipeCardAtIndex index: UInt, inDirection direction: SwipeResultDirection) -> Bool
-    func koloda(_ koloda: KolodaView, didSwipeCardAtIndex index: UInt, inDirection direction: SwipeResultDirection)
+    func koloda(_ koloda: KolodaView, allowedDirectionsForIndex index: Int) -> [SwipeResultDirection]
+    func koloda(_ koloda: KolodaView, shouldSwipeCardAtIndex index: Int, inDirection direction: SwipeResultDirection) -> Bool
+    func koloda(_ koloda: KolodaView, didSwipeCardAtIndex index: Int, inDirection direction: SwipeResultDirection)
     func kolodaDidRunOutOfCards(_ koloda: KolodaView)
-    func koloda(_ koloda: KolodaView, didSelectCardAtIndex index: UInt)
+    func koloda(_ koloda: KolodaView, didSelectCardAtIndex index: Int)
     func kolodaShouldApplyAppearAnimation(_ koloda: KolodaView) -> Bool
     func kolodaShouldMoveBackgroundCard(_ koloda: KolodaView) -> Bool
     func kolodaShouldTransparentizeNextCard(_ koloda: KolodaView) -> Bool
     func koloda(_ koloda: KolodaView, draggedCardWithPercentage finishPercentage: CGFloat, inDirection direction: SwipeResultDirection)
     func kolodaDidResetCard(_ koloda: KolodaView)
     func kolodaSwipeThresholdRatioMargin(_ koloda: KolodaView) -> CGFloat?
-    func koloda(_ koloda: KolodaView, didShowCardAtIndex index: UInt)
-    func koloda(_ koloda: KolodaView, shouldDragCardAtIndex index: UInt ) -> Bool
+    func koloda(_ koloda: KolodaView, didShowCardAtIndex index: Int)
+    func koloda(_ koloda: KolodaView, shouldDragCardAtIndex index: Int ) -> Bool
     
 }
 
 public extension KolodaViewDelegate {
-    func koloda(_ koloda: KolodaView, shouldSwipeCardAtIndex index: UInt, inDirection direction: SwipeResultDirection) -> Bool { return true }
+    func koloda(_ koloda: KolodaView, shouldSwipeCardAtIndex index: Int, inDirection direction: SwipeResultDirection) -> Bool { return true }
     
-    func koloda(_ koloda: KolodaView, allowedDirectionsForIndex index: UInt) -> [SwipeResultDirection] { return [.Left, .Right] }
-    func koloda(_ koloda: KolodaView, didSwipeCardAtIndex index: UInt, inDirection direction: SwipeResultDirection) {}
+    func koloda(_ koloda: KolodaView, allowedDirectionsForIndex index: Int) -> [SwipeResultDirection] { return [.Left, .Right] }
+    func koloda(_ koloda: KolodaView, didSwipeCardAtIndex index: Int, inDirection direction: SwipeResultDirection) {}
     func kolodaDidRunOutOfCards(_ koloda: KolodaView) {}
-    func koloda(_ koloda: KolodaView, didSelectCardAtIndex index: UInt) {}
+    func koloda(_ koloda: KolodaView, didSelectCardAtIndex index: Int) {}
     func kolodaShouldApplyAppearAnimation(_ koloda: KolodaView) -> Bool { return true }
     func kolodaShouldMoveBackgroundCard(_ koloda: KolodaView) -> Bool { return true }
     func kolodaShouldTransparentizeNextCard(_ koloda: KolodaView) -> Bool { return true }
     func koloda(_ koloda: KolodaView, draggedCardWithPercentage finishPercentage: CGFloat, inDirection direction: SwipeResultDirection) {}
     func kolodaDidResetCard(_ koloda: KolodaView) {}
     func kolodaSwipeThresholdRatioMargin(_ koloda: KolodaView) -> CGFloat? { return nil}
-    func koloda(_ koloda: KolodaView, didShowCardAtIndex index: UInt) {}
-    func koloda(_ koloda: KolodaView, shouldDragCardAtIndex index: UInt ) -> Bool { return true }
+    func koloda(_ koloda: KolodaView, didShowCardAtIndex index: Int) {}
+    func koloda(_ koloda: KolodaView, shouldDragCardAtIndex index: Int ) -> Bool { return true }
     
 }
 
@@ -122,7 +122,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
                 let countOfNeededCards = min(countOfVisibleCards, countOfCards - currentCardIndex)
                 
                 for index in 0..<countOfNeededCards {
-                    let actualIndex = UInt(index + currentCardIndex)
+                    let actualIndex = index + currentCardIndex
                     let nextCardView = createCardAtIndex(actualIndex)
                     let isTop = index == 0
                     nextCardView.isUserInteractionEnabled = isTop
@@ -133,23 +133,23 @@ open class KolodaView: UIView, DraggableCardDelegate {
                     visibleCards.append(nextCardView)
                     isTop ? addSubview(nextCardView) : insertSubview(nextCardView, belowSubview: visibleCards[index - 1])
                 }
-                self.delegate?.koloda(self, didShowCardAtIndex: UInt(currentCardIndex))
+                self.delegate?.koloda(self, didShowCardAtIndex: currentCardIndex)
             }
         }
     }
     
     public func layoutDeck() {
         for (index, card) in visibleCards.enumerated() {
-            layoutCard(card, atIndex: UInt(index))
+            layoutCard(card, atIndex: index)
         }
     }
     
-    private func layoutCard(_ card: DraggableCardView, atIndex index: UInt) {
+    private func layoutCard(_ card: DraggableCardView, atIndex index: Int) {
         if index == 0 {
             card.layer.transform = CATransform3DIdentity
             card.frame = frameForTopCard()
         } else {
-            let cardParameters = backgroundCardParametersForFrame(frameForCardAtIndex(UInt(index)))
+            let cardParameters = backgroundCardParametersForFrame(frameForCardAtIndex(index))
             let scale = cardParameters.scale
             card.layer.transform = CATransform3DScale(CATransform3DIdentity, scale.width, scale.height, 1.0)
             card.frame = cardParameters.frame
@@ -157,7 +157,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
     }
     
     //MARK: Frames
-    open func frameForCardAtIndex(_ index: UInt) -> CGRect {
+    open func frameForCardAtIndex(_ index: Int) -> CGRect {
         let bottomOffset:CGFloat = 0
         let topOffset = defaultBackgroundCardsTopMargin * CGFloat(countOfVisibleCards - 1)
         let scalePercent = defaultBackgroundCardsScalePercent
@@ -190,8 +190,8 @@ open class KolodaView: UIView, DraggableCardDelegate {
     internal func moveOtherCardsWithPercentage(_ percentage: CGFloat) {
         if visibleCards.count > 1 {
             for index in 1..<visibleCards.count {
-                let previousCardFrame = frameForCardAtIndex(UInt(index - 1))
-                var frame = frameForCardAtIndex(UInt(index))
+                let previousCardFrame = frameForCardAtIndex(index - 1)
+                var frame = frameForCardAtIndex(index)
                 let fraction = percentage / 100
                 
                 let distanceToMoveY: CGFloat = (frame.origin.y - previousCardFrame.origin.y) * fraction
@@ -257,12 +257,12 @@ open class KolodaView: UIView, DraggableCardDelegate {
     }
     
     func card(_ card: DraggableCardView, shouldSwipeInDirection direction: SwipeResultDirection) -> Bool {
-        return delegate?.koloda(self, shouldSwipeCardAtIndex: UInt(self.currentCardIndex), inDirection: direction) ?? true
+        return delegate?.koloda(self, shouldSwipeCardAtIndex: self.currentCardIndex, inDirection: direction) ?? true
     }
     
     func card(cardAllowedDirections card: DraggableCardView) -> [SwipeResultDirection] {
         let index = currentCardIndex + visibleCards.index(of: card)!
-        return delegate?.koloda(self, allowedDirectionsForIndex: UInt(index)) ?? [.Left, .Right]
+        return delegate?.koloda(self, allowedDirectionsForIndex: index) ?? [.Left, .Right]
     }
     
     func card(_ card: DraggableCardView, wasSwipedInDirection direction: SwipeResultDirection) {
@@ -297,7 +297,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
         guard let visibleIndex = visibleCards.index(of: card) else { return }
         
         let index = currentCardIndex + visibleIndex
-        delegate?.koloda(self, didSelectCardAtIndex: UInt(index))
+        delegate?.koloda(self, didSelectCardAtIndex: index)
     }
     
     func card(cardSwipeThresholdRatioMargin card: DraggableCardView) -> CGFloat? {
@@ -308,7 +308,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
         guard let visibleIndex = visibleCards.index(of: card) else { return true}
         
         let index = currentCardIndex + visibleIndex
-        return delegate?.koloda(self, shouldDragCardAtIndex: UInt(index)) ?? true
+        return delegate?.koloda(self, shouldDragCardAtIndex: index) ?? true
     }
     
     //MARK: Private
@@ -341,12 +341,12 @@ open class KolodaView: UIView, DraggableCardDelegate {
                 
                 _self.visibleCards.last?.isHidden = false
                 _self.animating = false
-                _self.delegate?.koloda(_self, didSwipeCardAtIndex: UInt(_self.currentCardIndex - 1), inDirection: direction)
-                _self.delegate?.koloda(_self, didShowCardAtIndex: UInt(_self.currentCardIndex))
+                _self.delegate?.koloda(_self, didSwipeCardAtIndex: _self.currentCardIndex - 1, inDirection: direction)
+                _self.delegate?.koloda(_self, didShowCardAtIndex: _self.currentCardIndex)
             }
         } else {
             animating = false
-            delegate?.koloda(self, didSwipeCardAtIndex: UInt(self.currentCardIndex - 1), inDirection: direction)
+            delegate?.koloda(self, didSwipeCardAtIndex: self.currentCardIndex - 1, inDirection: direction)
             delegate?.kolodaDidRunOutOfCards(self)
         }
     }
@@ -356,8 +356,8 @@ open class KolodaView: UIView, DraggableCardDelegate {
             return
         }
         
-        let cardParameters = backgroundCardParametersForFrame(frameForCardAtIndex(UInt(visibleCards.count)))
-        let lastCard = createCardAtIndex(UInt(currentCardIndex + countOfVisibleCards - 1), frame: cardParameters.frame)
+        let cardParameters = backgroundCardParametersForFrame(frameForCardAtIndex(visibleCards.count))
+        let lastCard = createCardAtIndex(currentCardIndex + countOfVisibleCards - 1, frame: cardParameters.frame)
         
         let scale = cardParameters.scale
         lastCard.layer.transform = CATransform3DScale(CATransform3DIdentity, scale.width, scale.height, 1)
@@ -377,7 +377,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
             currentCard.removeAnimations()
             
             currentCard.isUserInteractionEnabled = index == 0
-            let cardParameters = backgroundCardParametersForFrame(frameForCardAtIndex(UInt(index)))
+            let cardParameters = backgroundCardParametersForFrame(frameForCardAtIndex(index))
             var animationCompletion: ((Bool) -> Void)? = nil
             if index != 0 {
                 if shouldTransparentizeNextCard {
@@ -417,7 +417,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
             currentCardIndex -= 1
             
             if dataSource != nil {
-                let firstCardView = createCardAtIndex(UInt(currentCardIndex), frame: frameForTopCard())
+                let firstCardView = createCardAtIndex(currentCardIndex, frame: frameForTopCard())
                 
                 if shouldTransparentizeNextCard {
                     firstCardView.alpha = alphaValueTransparent
@@ -434,7 +434,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
                     }
                     
                     _self.animating = false
-                    _self.delegate?.koloda(_self, didShowCardAtIndex: UInt(_self.currentCardIndex))
+                    _self.delegate?.koloda(_self, didShowCardAtIndex: _self.currentCardIndex)
                     })
             }
             
@@ -444,7 +444,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
                 }
                 card.isUserInteractionEnabled = false
                 
-                let cardParameters = backgroundCardParametersForFrame(frameForCardAtIndex(UInt(index + 1)))
+                let cardParameters = backgroundCardParametersForFrame(frameForCardAtIndex(index + 1))
                 animator.applyScaleAnimation(
                     card,
                     scale: cardParameters.scale,
@@ -464,11 +464,11 @@ open class KolodaView: UIView, DraggableCardDelegate {
             
             for index in startIndex...endIndex {
                 let nextCardView = generateCard(frameForTopCard())
-                layoutCard(nextCardView, atIndex: UInt(index))
+                layoutCard(nextCardView, atIndex: index)
                 nextCardView.alpha = shouldTransparentizeNextCard ? alphaValueSemiTransparent : alphaValueOpaque
                 
                 visibleCards.append(nextCardView)
-                configureCard(nextCardView, atIndex: UInt(currentCardIndex + index))
+                configureCard(nextCardView, atIndex: currentCardIndex + index)
                 if index > 0 {
                     insertSubview(nextCardView, belowSubview: visibleCards[index - 1])
                 } else {
@@ -481,7 +481,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
     private func reconfigureCards() {
         if dataSource != nil {
             for (index, card) in visibleCards.enumerated() {
-                let actualIndex = UInt(currentCardIndex + index)
+                let actualIndex = currentCardIndex + index
                 configureCard(card, atIndex: actualIndex)
             }
         }
@@ -520,7 +520,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
     
     public func swipe(_ direction: SwipeResultDirection) {
         
-        let validDirection = delegate?.koloda(self, allowedDirectionsForIndex: UInt(currentCardIndex)).contains(direction) ?? true
+        let validDirection = delegate?.koloda(self, allowedDirectionsForIndex: currentCardIndex).contains(direction) ?? true
         guard validDirection else { return }
         
         if !animating {
@@ -563,7 +563,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
     private func insertVisibleCardsWithIndexes(_ visibleIndexes: [Int]) -> [DraggableCardView] {
         var insertedCards: [DraggableCardView] = []
         visibleIndexes.forEach { insertionIndex in
-            let card = createCardAtIndex(UInt(insertionIndex))
+            let card = createCardAtIndex(insertionIndex)
             let visibleCardIndex = insertionIndex - currentCardIndex
             visibleCards.insert(card, at: visibleCardIndex)
             if visibleCardIndex == 0 {
@@ -575,7 +575,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
                 card.alpha = shouldTransparentizeNextCard ? alphaValueSemiTransparent : alphaValueOpaque
                 insertSubview(card, belowSubview: visibleCards[visibleCardIndex - 1])
             }
-            layoutCard(card, atIndex: UInt(visibleCardIndex))
+            layoutCard(card, atIndex: visibleCardIndex)
             insertedCards.append(card)
         }
         
@@ -684,7 +684,7 @@ open class KolodaView: UIView, DraggableCardDelegate {
             let visibleCardIndex = index - currentCardIndex
             if visibleCards.count > visibleCardIndex {
                 let card = visibleCards[visibleCardIndex]
-                configureCard(card, atIndex: UInt(index))
+                configureCard(card, atIndex: index)
             }
         }
     }
